Separating out models to their individual files and updating CityWeatherView to have another Reusable view.
